Move codec creation from `configmaps.go` to `kube_apiserver.go`
Since this is no longer restricted to content in `configmaps.go` only
